UI - cluster item classes : renaming concrete cluster item files
