+++ src/ui/simulator/toolbox/components/htmllistbox/item/thermal-cluster-item.h
